changes for This build :i found bugs in part 1 so i have fixed them. also, changed that all the boards game will start at 1 position not at 0 (this is becuase part two of the ptoject) and also fixed little things in part 2 -> still need to check part two. but i have tested part 1 and it seems to work.
but there are major changes in part 1 which had bugs, like didnt the main function retunes null and valid position is now by refernce to the current position..
------------------------------------------------
VER 2.0.
--------
part two of the project checked and works good, with the input that also added.
also adding free functions for the list and board.